Replaced legacy libraries with NuGet versions and upgraded project to .NET 8

+++ DS_Map/DSUtils/ARM9.cs
@@ -56,7 +56,7 @@ namespace DSPRE {
             return DSUtils.ReadFromFile(RomInfo.arm9Path, startOffset, 1)[0];
         }
         public static void WriteByte(byte value, uint destOffset) {
-            DSUtils.WriteToFile(RomInfo.arm9Path, BitConverter.GetBytes(value), destOffset, 0);
+            DSUtils.WriteToFile(RomInfo.arm9Path, BitConverter.GetBytes((ushort)value), destOffset, 0);
         }
     }
 }
